+++ src/Symfony/Bridge/Twig/Tests/Extension/TranslationExtensionTest.php
@@ -268,6 +268,282 @@ class TranslationExtensionTest extends TestCase
         $this->assertEquals('foo (custom)foo (foo)foo (custom)', trim($template->render(['foo_domain' => 'foo', 'custom_domain' => 'custom'])));
     }
 
+    public function testForSelfDefaultTranslationDomain()
+    {
+        $templates = [
+            'index' => '
+                {%- trans_default_domain "foo" for _self -%}
+                1. {{ "key"|trans -}}{{" "-}}
+                2. {% trans %}key{% endtrans %}{{" "-}}
+                {%- embed "embedded.html.twig" -%}
+                    {%- block content -%}
+                    3. {{ "key"|trans }}{{" "-}}
+                    {%- endblock -%}
+                {%- endembed %}
+            ',
+
+            // another template, with its own "for _self" that applies to its own trans calls only
+            'embedded.html.twig' => '
+                {%- trans_default_domain "bar" for _self -%}
+                {%- block content "" -%}
+                4. {{ "key"|trans -}}
+            ',
+        ];
+
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['key' => 'key-messages'], 'en');
+        $translator->addResource('array', ['key' => 'key-foo'], 'en', 'foo');
+        $translator->addResource('array', ['key' => 'key-bar'], 'en', 'bar');
+
+        $template = $this->getTemplate($templates, $translator);
+
+        $this->assertEquals('1. key-foo 2. key-foo 3. key-foo 4. key-bar', trim($template->render([])));
+    }
+
+    public function testForSelfDefaultTranslationDomainWithNestedEmbed()
+    {
+        $templates = [
+            'index' => '
+                {%- trans_default_domain "foo" for _self -%}
+                {%- embed "outer_embed.html.twig" -%}
+                    {%- block content -%}
+                        1. {{ "key"|trans -}}{{" "-}}
+                        {%- embed "inner_embed.html.twig" -%}
+                            {%- block inner -%}
+                            2. {{ "key"|trans }}{{" "-}}
+                            {%- endblock -%}
+                        {%- endembed -%}
+                    {%- endblock -%}
+                {%- endembed -%}
+            ',
+
+            'outer_embed.html.twig' => '
+                {%- block content "" -%}
+            ',
+
+            // the trans call below is written in another template and must not be affected
+            'inner_embed.html.twig' => '
+                {%- block inner "" -%}
+                3. {{ "key"|trans -}}
+            ',
+        ];
+
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['key' => 'key-messages'], 'en');
+        $translator->addResource('array', ['key' => 'key-foo'], 'en', 'foo');
+
+        $template = $this->getTemplate($templates, $translator);
+
+        $this->assertEquals('1. key-foo 2. key-foo 3. key-messages', trim($template->render([])));
+    }
+
+    public function testForSelfDefaultTranslationDomainAppliesToMacros()
+    {
+        $template = '
+            {%- trans_default_domain "foo" for _self -%}
+            {{- _self.one() }} {{ _self.two() -}}
+            {%- macro one() -%}
+                1. {{ "key"|trans -}}
+            {%- endmacro -%}
+            {%- macro two() -%}
+                2. {{ "key"|trans -}}
+            {%- endmacro -%}
+        ';
+
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['key' => 'key-messages'], 'en');
+        $translator->addResource('array', ['key' => 'key-foo'], 'en', 'foo');
+
+        $this->assertEquals('1. key-foo 2. key-foo', trim($this->getTemplate($template, $translator)->render([])));
+    }
+
+    public function testForSelfDefaultTranslationDomainAppliesToInheritedBlocks()
+    {
+        $templates = [
+            'index' => '
+                {%- extends "layout.html.twig" -%}
+                {%- trans_default_domain "foo" for _self -%}
+                {%- block content -%}{{ "key"|trans }}{%- endblock -%}
+            ',
+            'layout.html.twig' => '{% block content "" %}',
+        ];
+
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['key' => 'key-messages'], 'en');
+        $translator->addResource('array', ['key' => 'key-foo'], 'en', 'foo');
+
+        $this->assertEquals('key-foo', trim($this->getTemplate($templates, $translator)->render([])));
+    }
+
+    public function testForSelfDefaultTranslationDomainDoesNotLeakIntoIncludedTemplates()
+    {
+        $templates = [
+            'index' => '{% trans_default_domain "foo" for _self %}{% include "included.html.twig" %}',
+            'included.html.twig' => '{{ "key"|trans }}',
+        ];
+
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['key' => 'key-messages'], 'en');
+        $translator->addResource('array', ['key' => 'key-foo'], 'en', 'foo');
+
+        $this->assertEquals('key-messages', trim($this->getTemplate($templates, $translator)->render([])));
+    }
+
+    public function testForSelfDefaultTranslationDomainPrecedence()
+    {
+        $template = '
+            {%- trans_default_domain "foo" for _self -%}
+            1. {{ "key"|trans -}}{{" "-}}
+            2. {{ "key"|trans({}, "explicit") -}}{{" "-}}
+            {%- trans_default_domain "bar" -%}
+            3. {{ "key"|trans -}}{{" "-}}
+            4. {{ "key"|trans({}, "explicit") -}}{{" "-}}
+            {%- trans_default_domain "baz" -%}
+            5. {{ "key"|trans -}}{{" "-}}
+            6. {{ "key"|trans({}, "explicit") -}}
+        ';
+
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['key' => 'key-messages'], 'en');
+        $translator->addResource('array', ['key' => 'key-foo'], 'en', 'foo');
+        $translator->addResource('array', ['key' => 'key-bar'], 'en', 'bar');
+        $translator->addResource('array', ['key' => 'key-baz'], 'en', 'baz');
+        $translator->addResource('array', ['key' => 'key-explicit'], 'en', 'explicit');
+
+        // a scoped trans_default_domain wins over "for _self", an explicit domain wins over both
+        $this->assertEquals('1. key-foo 2. key-explicit 3. key-bar 4. key-explicit 5. key-baz 6. key-explicit', trim($this->getTemplate($template, $translator)->render([])));
+    }
+
+    public function testForSelfDefaultTranslationDomainCanBeOverriddenInsideAnEmbed()
+    {
+        $templates = [
+            'index' => '
+                {%- trans_default_domain "foo" for _self -%}
+                {%- embed "embedded.html.twig" -%}
+                    {%- block content -%}
+                        {%- trans_default_domain "bar" -%}
+                        {{- "key"|trans -}}
+                    {%- endblock -%}
+                {%- endembed -%}
+            ',
+            'embedded.html.twig' => '{% block content "" %}',
+        ];
+
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['key' => 'key-foo'], 'en', 'foo');
+        $translator->addResource('array', ['key' => 'key-bar'], 'en', 'bar');
+
+        $this->assertEquals('key-bar', trim($this->getTemplate($templates, $translator)->render([])));
+    }
+
+    public function testForSelfDefaultTranslationDomainAppliesToCallsBeforeDeclaration()
+    {
+        $template = '
+            1. {{ "key"|trans -}}{{" "-}}
+            {%- trans_default_domain "foo" for _self -%}
+            2. {{ "key"|trans -}}
+        ';
+
+        $translator = new Translator('en');
+        $translator->addLoader('array', new ArrayLoader());
+        $translator->addResource('array', ['key' => 'key-messages'], 'en');
+        $translator->addResource('array', ['key' => 'key-foo'], 'en', 'foo');
+
+        $this->assertEquals('1. key-foo 2. key-foo', trim($this->getTemplate($template, $translator)->render([])));
+    }
+
+    public function testForSelfDefaultTranslationDomainIsExtracted()
+    {
+        $templates = [
+            'index' => '
+                {%- trans_default_domain "foo" for _self -%}
+                {{ "outer"|trans }}
+                {%- embed "embedded.html.twig" -%}
+                    {%- block content -%}{{ "embedded"|trans }}{%- endblock -%}
+                {%- endembed -%}
+            ',
+            'embedded.html.twig' => '{% block content "" %}',
+        ];
+
+        $twig = new Environment(new TwigArrayLoader($templates), ['debug' => true, 'cache' => false]);
+        $twig->addExtension($extension = new TranslationExtension(new Translator('en')));
+
+        $visitor = $extension->getTranslationNodeVisitor();
+        $visitor->enable();
+        $twig->parse($twig->tokenize($twig->getLoader()->getSourceContext('index')));
+
+        $this->assertEquals([['embedded', 'foo'], ['outer', 'foo']], $visitor->getMessages());
+    }
+
+    public function testForSelfWithNonConstantDomainThrows()
+    {
+        $this->expectException(SyntaxError::class);
+        $this->expectExceptionMessage('The "for _self" modifier of the "trans_default_domain" tag requires a constant domain');
+        $this->getTemplate('{% trans_default_domain domain for _self %}');
+    }
+
+    public function testForSelfInsideBlockThrows()
+    {
+        $this->expectException(SyntaxError::class);
+        $this->expectExceptionMessage('The "trans_default_domain" tag cannot be used with "for _self" inside a block');
+        $this->getTemplate('{% block content %}{% trans_default_domain "foo" for _self %}{% endblock %}');
+    }
+
+    public function testForSelfDeclaredTwiceThrows()
+    {
+        $this->expectException(SyntaxError::class);
+        $this->expectExceptionMessage('The "trans_default_domain" tag can be used with "for _self" only once per template');
+        $this->getTemplate('{% trans_default_domain "foo" for _self %}{% trans_default_domain "bar" for _self %}');
+    }
+
+    public function testAFailingTemplateDoesNotLeakItsDomainIntoTheNextOne()
+    {
+        $translator = new Translator('en');
+        $loader = new TwigArrayLoader([
+            'broken' => '{% block content %}{% trans_default_domain "foo" for _self %}{% endblock %}',
+            'other' => '{{ "key"|trans }}',
+        ]);
+        $twig = new Environment($loader, ['debug' => true, 'cache' => false]);
+        $twig->addExtension(new TranslationExtension($translator));
+
+        try {
+            $twig->load('broken');
+            $this->fail('The template is expected to fail.');
+        } catch (SyntaxError) {
+        }
+
+        $source = $twig->compile($twig->parse($twig->tokenize($twig->getLoader()->getSourceContext('other'))));
+
+        $this->assertStringNotContainsString('foo', $source);
+    }
+
+    public function testForSelfInsideEmbedThrows()
+    {
+        $this->expectException(SyntaxError::class);
+        $this->expectExceptionMessage('The "trans_default_domain" tag cannot be used with "for _self" inside an "embed" tag');
+        $this->getTemplate([
+            'index' => '{% embed "layout.html.twig" %}{% trans_default_domain "foo" for _self %}{% block content %}{% endblock %}{% endembed %}',
+            'layout.html.twig' => '{% block content "" %}',
+        ]);
+    }
+
+    public function testForSelfAfterEmbedThrows()
+    {
+        $this->expectException(SyntaxError::class);
+        $this->expectExceptionMessage('The "trans_default_domain" tag must be used with "for _self" before any "embed" tag');
+        $this->getTemplate([
+            'index' => '{% embed "layout.html.twig" %}{% block content %}{% endblock %}{% endembed %}{% trans_default_domain "foo" for _self %}',
+            'layout.html.twig' => '{% block content "" %}',
+        ]);
+    }
+
     private function getTemplate($template, ?TranslatorInterface $translator = null): TemplateWrapper
     {
         $translator ??= new Translator('en');